cat_automation/requirements.txt
anthropic>=0.25.0
requests>=2.31.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
edge-tts>=6.1.0
